Refactor debug menu packages
