+++ modules/debug_menu/src/main/kotlin/kekmech/ru/debug_menu/initializer/DebugMenuInitializer.kt
@@ -7,7 +7,6 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import kekmech.ru.common_app_lifecycle.AppLifecycleObserver
-import kekmech.ru.debug_menu.presentation.DebugMenuService
 import timber.log.Timber
 
 internal class DebugMenuInitializer : AppLifecycleObserver {
